refactor _sdkver to wasm and wasm_min package
