+++ wasm/wallet.go
@@ -1,4 +1,4 @@
-package main
+package wasm
 
 import (
 	"encoding/json"
