+++ wasm_min/proxy_min.go
@@ -1,4 +1,4 @@
-package main
+package wasm_min
 
 import (
 	"fmt"
@@ -977,48 +977,12 @@ func main() {
 	js.Global().Set("Encrypt", js.FuncOf(Encrypt))
 	js.Global().Set("Decrypt", js.FuncOf(Decrypt))
 
-	// zboxsdk.go
+	// zboxsdk_min.go
 	js.Global().Set("InitAuthTicket", js.FuncOf(InitAuthTicket))
-	js.Global().Set("InitZCNSDK", js.FuncOf(InitZCNSDK))
-	js.Global().Set("ZBOXSetLogLevel", js.FuncOf(ZBOXSetLogLevel))
-	js.Global().Set("ZBOXSetLogFile", js.FuncOf(ZBOXSetLogFile))
-	js.Global().Set("GetNetwork", js.FuncOf(GetNetwork))
-	js.Global().Set("SetMaxTxnQuery", js.FuncOf(SetMaxTxnQuery))
-	js.Global().Set("SetQuerySleepTime", js.FuncOf(SetQuerySleepTime))
-	js.Global().Set("SetMinSubmit", js.FuncOf(SetMinSubmit))
-	js.Global().Set("SetMinConfirmation", js.FuncOf(SetMinConfirmation))
-	js.Global().Set("ZBOXSetNetwork", js.FuncOf(ZBOXSetNetwork))
-	js.Global().Set("CreateReadPool", js.FuncOf(CreateReadPool))
-	js.Global().Set("AllocFilter", js.FuncOf(AllocFilter))
-	js.Global().Set("ZBOXGetReadPoolInfo", js.FuncOf(ZBOXGetReadPoolInfo))
-	js.Global().Set("ReadPoolLock", js.FuncOf(ReadPoolLock))
-	js.Global().Set("ReadPoolUnlock", js.FuncOf(ReadPoolUnlock))
-	js.Global().Set("ZBOXGetStakePoolInfo", js.FuncOf(ZBOXGetStakePoolInfo))
-	js.Global().Set("ZBOXGetStakePoolUserInfo", js.FuncOf(ZBOXGetStakePoolUserInfo))
-	js.Global().Set("StakePoolLock", js.FuncOf(StakePoolLock))
-	js.Global().Set("StakePoolUnlock", js.FuncOf(StakePoolUnlock))
-	js.Global().Set("StakePoolPayInterests", js.FuncOf(StakePoolPayInterests))
-	js.Global().Set("ZBOXGetWritePoolInfo", js.FuncOf(ZBOXGetWritePoolInfo))
-	js.Global().Set("WritePoolLock", js.FuncOf(WritePoolLock))
-	js.Global().Set("WritePoolUnlock", js.FuncOf(WritePoolUnlock))
-	js.Global().Set("ZBOXGetChallengePoolInfo", js.FuncOf(ZBOXGetChallengePoolInfo))
-	js.Global().Set("ZBOXGetStorageSCConfig", js.FuncOf(ZBOXGetStorageSCConfig))
-	js.Global().Set("ZBOXGetBlobbers", js.FuncOf(ZBOXGetBlobbers))
-	js.Global().Set("ZBOXGetBlobber", js.FuncOf(ZBOXGetBlobber))
 	js.Global().Set("ZBOXGetClientEncryptedPublicKey", js.FuncOf(ZBOXGetClientEncryptedPublicKey))
-	js.Global().Set("GetAllocationFromAuthTicket", js.FuncOf(GetAllocationFromAuthTicket))
 	js.Global().Set("ZBOXGetAllocation", js.FuncOf(ZBOXGetAllocation))
-	js.Global().Set("ZBOXGetAllocations", js.FuncOf(ZBOXGetAllocations))
-	js.Global().Set("GetAllocationsForClient", js.FuncOf(GetAllocationsForClient))
-	js.Global().Set("CreateAllocation", js.FuncOf(CreateAllocation))
-	js.Global().Set("CreateAllocationForOwner", js.FuncOf(CreateAllocationForOwner))
-	js.Global().Set("UpdateAllocation", js.FuncOf(UpdateAllocation))
-	js.Global().Set("FinalizeAllocation", js.FuncOf(FinalizeAllocation))
-	js.Global().Set("CancelAllocation", js.FuncOf(CancelAllocation))
-	js.Global().Set("UpdateBlobberSettings", js.FuncOf(UpdateBlobberSettings))
-	js.Global().Set("CommitToFabric", js.FuncOf(CommitToFabric))
-	js.Global().Set("GetAllocationMinLock", js.FuncOf(GetAllocationMinLock))
 	js.Global().Set("SetNumBlockDownloads", js.FuncOf(SetNumBlockDownloads))
-
+	js.Global().Set("ZBOXGetAllocations", js.FuncOf(ZBOXGetAllocations))
+	js.Global().Set("GetAllocationFromAuthTicket", js.FuncOf(GetAllocationFromAuthTicket))
 	<-c
 }
